Finish step 2: Poisson Lee-Carter fit and kt/e65 forecasting
- poisson_lee_carter.py: complete the Newton-Raphson-calibrated Poisson
  MLE kt fit (initial_guess, update_ax/bx/kt, fit_poisson_lc,
  fit_poisson_lc_all), replacing the previous stub.
- lc_outputs.py: split out of lee_carter.py as the shared LC
  diagnostics/A-E/e-at-age layer (reconstruct_lc, build_lc_tables,
  plot_lc_parameters, ae_metric, ae_heatmap, ae_summary,
  ex_progression_plot) so both the SVD and Poisson fits reuse it.
- forecasting.py: add a 95% CI band to the kt drift forecast
  (forecast_kt/plot_kt_forecast) and propagate it through to an e65
  forecast fan chart (project_ln_mx -> build_forecast_life_table ->
  forecast_ex_fan -> plot_ex_fan_chart).
- step2_all_dataset_model.py: wire the kt forecast and e65 fan chart
  through for both the SVD and Poisson Lee-Carter fits.
- config.yaml: resolve the state/cause/race subsection scope decision
  and add the API credibility gate spec.
- eda.py/load_filter.py: minor groupby deprecation and dtype fixes.
- CLAUDE.md: update to reflect current state, correcting the stale
  poisson_lee_carter.py stub note.

+++ src/actuarial/load_filter.py
@@ -13,7 +13,7 @@ def load_filter(
     df = pd.read_csv(path)
     validation_df = df.copy()
 
-    validation_df["Year Code"] = pd.to_numeric(validation_df["Year Code"])
+    validation_df["Year Code"] = pd.to_numeric(validation_df["Year Code"]).astype(int)
     validation_df["Deaths"] = pd.to_numeric(
         validation_df["Deaths"],
         errors='coerce',
